Implementation of a very basic and rough API
Just an MVP/prototype for now, this basic API has the necessary endpoints to create applications, list them, get an application by id, and update the status of an application by id.

Added a README with some documentation on how to setup the local service.

Docker and docker-compose are also functional.

Again, this is to start with something. The declaration of fields in the `CrimeApplication` model is by no means comprehensive and is lacking proper indexes (still playing and reading about it).