Add kubernetes manifests
The deployment requires the dynamodb endpoint or credentials coming from terraform `env`. We don't have yet a dynamodb resource in the namespace so that will be added when we have it.

The rest is fairly standard and almost identical to our other services.

In the `secrets.yml` file are the http auth credentials.