Avoid `MissingRangeKey` when using `#find`
It turns out the `#find` method from dynamoid forces you to pass the range when a range_key is declared in the table, which is annoying and really pointless I suppose when we have a hash key on the UUID and these are unique.

Create a convenience method wrapper for when we don't pass the `range_key` option. Still uses `super` if this is passed.