Implement scope by office code in list endpoint
Apply requires filtering by office code. All requests to GET /applications will include, in addition to the `status` of the application, the `office_code` of the provider that submitted it.

I've done the filtering with a scope based on the JSON stored in the `application` DB field.
This will unlock additional functionality on Apply side.

I'm sure there are better ways to do this, as this probably is slow, specially without an index. Please feel free to change this to virtual attributes or whatever if that's better.