Refactor list applications sorting and pagination (#32)
* Refactor list applications sorting and pagination

Some due refactor of the sorting. Made it behave similar to the existing Pagination so it can use `apply_to_scope`.

It now support passing an `order` to choose the column to sort by (previously we were always hardcoding this to either submitted_at or returned_at). Fixes a bug on Apply dashboard.

Reorganised the helpers.

Ensure the tests are exercising the different sort columns and directions.

Made it `asc` and `ascending`, and `desc` and `descending` aliases.

NOTE 1: there are no functional changes, everything will keep working as it is (no changes needed on Apply and I suspect none on Review either). Same query params on the outside.

NOTE 2: I'm not really happy with the naming of the query params for sorting. They seem confusing to me. Currently we are using `order` for the column to sort by, and `sort` for the direction.
I've kept these on the outside interface so no changes to consumers are required, however, I would like to propose a different, more explicit naming:

- sort_by: the column to use for the sorting
- direction (or maybe we can use `order` for this): the direction, asc / desc (ascending / descending accepted).

NOTE 3: I think there is a confusing behaviour in the datastore with `submitted_at` dates. At the moment, if I'm correct, the datastore is not using the JSON payload `submitted_at` attribute and instead it is assigning the current `CURRENT_TIMESTAMP` on inserts.
This, mostly works, however, it is confusing to then send the attribute in the payload, and we may consider removing it like we did with `status`.

However, the `date_stamp` is sometimes set to the exact submission date, on Apply, under certain case types. This means with the current code on datastore, there is an edge case where a date stamp and a submitted at might not match once the application is created in the datastore and returned to the consumer (server time drift).

To be discussed?

* Rename query params for sorting

As agreed with other devs.

Apply who is the only consumer currently performing sorting, will need to update the query params in use.