Cleanup dynamodb (#35)
This removes all code related to dynamodb and API v1 endpoints.

Keeping postgres API as v2 for the time being. At some point we can decide to rename to v1 and update consumers of the API to point (again) to v1.

Will be removing the dynamodb cloud resource from the namespace too.