CRIMAP-265 Supersede parent of resubmissions
On resubmission of a returned application, this new submission is technically a brand new application, but points to the returned one through an attribute `parent_id`.

With this information, when a resubmission (meaning  `parent_id` is not nil) is received and successfully validated and persisted, the “parent“ application is then marked as superseded, by changing its status.

Additionally in this PR some minor tweaks:

- shift the `submitted_at` attribute from the payload to the table column so both are always in sync.

- use an enum for the `status` attribute on the CrimeApplication records, so no longer need to have a separate presence/inclusion validation, and also we gain methods like `returned?` and scopes.