Prepare datastore for JWT auth
Merging this will not enable JWT or change any functionality on the datastore, as it is not enabled yet (the line that would otherwise enable this is commented out on purpose for the easy testing in localhost).

Once the `auth :jwt` is enabled, it will make use of a little middleware introduced in the `simple-jwt-auth` gem to perform verification of the bearer token.