First incarnation of SNS events publishing
We now have an initial infrastructure created in the namespace, with an SNS topic that receives notifications, and an SQS queue where these notifications are sent (for now no filter).

Idea being, datastore will publish a series of events to the SNS topic, and then one or more SQS subscribe to this topic and each receives the notifications that matters to them, based on an event filter.

Each event has a name. Only 2 are created in this PR as an initial example. Also each event can implement its own message/payload. For now both are the same, they publish the whole application JSON. But some events might require only a few attributes so this is possible.

I've left previous code for job processor with `shoryuken` but I don't think we need that anymore and is an overkill for our needs. The publishing to the topic is very fast so there is no need to enqueue this to be processed at a later time. It removes a lot of complexity.
Idea is to clean up shoryuken unused code in a future PR unless we see a reason to leave it.