CRIMAP-285 Configure Sentry exceptions reporting
Almost identical to the setup we did for Apply.

Added the `HostEnv` class so we can have 1 sentry project handling the exceptions of the different namespaces instead of having to create 1 project per namespace.

So for instance, right now we only have `staging` namespace, exceptions will be reported with `environment` tag set to `staging`.
Not to be confused with the Rails/Rack env (which is always production no matter the namespace).

For now this will catch any unhandled exception. Later on we can add Sentry reporting for some handled exceptions with additional metadata.
Example: https://github.com/ministryofjustice/laa-apply-for-criminal-legal-aid/blob/main/app/services/datastore/application_counters.rb#L24