Tweaks to SNS publishing (#52)
Moved the publishing of the events into the transaction as discussed.

Modified the message payload format and content. Only the ID of the application is now included in the payload. Also added the `event_name` to the message body as some subscribers might not get it from the message attributes.

Removed the subject, as it is not useful now that we have the event name in the message itself.