CRIMAP-289 Configure logging and filtering of PII
Logs to stdout so other systems can route these logs to final destinations for viewing and long-term archival. For example this is needed to see the logs in Kibana.

Also, added a few more PII parameters to the filter. These are also used by Sentry when reporting exceptions.

NOTE: not using `lograge` as we did on Apply because that gem is only for Rails/Rails API, but Datastore is Grape API. Using an alternative gem which is similar and also produce a `logstash` format.

It is easy to extend to log more stuff by implementing a `GrapeLogging::Loggers` class. For instance I've implemented one for logging the JWT issuer of the request:
`lib/grape_logging/loggers/jwt_issuer.rb`