CRIMAP-287 Decommission datastore-harness namespace (#55)
This `harness` namespace was created back when we had DynamoDB and we thought having a remote, "real" dynamodb might be easier than setting one locally or using docker.

It turned out this probably was never used. Running a local dynamodb is not that complicated after all.

Removing left overs. Will also raise a parallel PR to cloud-platforms to delete the resources and namespace.

I think Apply has some references to the harness (in README or env files), will take a look at that too and raise if needed a PR there.