Refactor and cleanup health check
The health check endpoint was a weird one. It was part of the API (with versioning in the path and all) but not really coupled to anything in particular to grant this level of coupling.

Some of it was left overs of our integration with DynamoDB.

Additionally, after the introduction of the JWT token, the health endpoint required also authentication, which seems to defeat the purpose.

I've refactor this into just a normal controller action, like the `/ping` action we have, so it can be called without auth and thus we can start using it also in the pods probes.

Also, report to sentry any failures on DB connectivity.