CRIMAP-38 Task to reassign apps to new office code
Not to be merged until Apply PR is merged and tested.
https://github.com/ministryofjustice/laa-apply-for-criminal-legal-aid/pull/292

This will reassign all applications in the datastore to an office belonging to the new provider we will be using on Apply staging after integrating with Portal staging.

Previous office code (`1A123B`) belongs to a mock provider, while we used SAMLmock on staging.

The only purpose of this is to keep going as if nothing happened, to not loose all the test applications we already have, needed for testing pagination, and sorting, etc.

To be able to do this more easily making use of active record, temporarily the `application` attribute has been removed from the `attr_readonly` declaration.

After this is merged, and the task is run, a follow-up PR will be raised to clean this up.