CRIMAP-288 Cleanup shoryuken code
Introduced in PR #40, this is almost a rollback (with some exceptions) of most of the code.

Originally introduced as part of the spike to send messages to an SQS queue, in reality a job processor was not needed, and introduces unnecessary complexity. Publishing a message takes milliseconds and is not a time-consuming job.

Also, we've now taken a different approach to queues and messaging. We are sending notifications to an SNS queue, and other subscribers (including SQS if needed) can subscribe to this SNS topic.
The AWS gem is enough for this task, no other gems are required.