CRIMAP-349 Perform authorisation per endpoint
The datastore currently implements authentication via JWT tokens but it doesn’t perform “authorisation“ to limit access of consumers to endpoints that are out of their business.

For instance, we would want MAAT adapter consumer to only have access to the endpoint(s) relevant to them, and to not have access to the create application endpoint, or the search applications endpoint.

In order to do that we have to “annotate“ each endpoint to declare which consumers (JWT issuer) are allowed to call those endpoints. If they are not allowed, return forbidden.

This is making use of the new Grape middleware implemented in the `SimpleJwtAuth` gem that handles this authorisation based on the annotation as a before filter.

NOTE: maat endpoint not yet annotated as it needs a refactor first and agreeing with them the issuer name.