CRIMAP-350 Rescope API back to v1
Get back to v1 in all API endpoints.

Also, maat endpoint has been rescoped to also be versioned, in the endpoint: `/api/:version/maat/applications/:usn`

As part of this shuffling around of files etc. I've also versioned all the entities as it is possible we may want to keep an endpoint in v1 but using a different entity, or the other way around.
Also makes easier to bump individual endpoints versions without disrupting existing ones.
This is only an internal refactor.

Added a rake task to list all existing endpoints with some details, in particular, the consumers authorised in each one: `rake grape:routes`.