Change the way SimpleJwtAuth sets the logger
Instead of `Rails.logger`, use the logger provided by Grape, which will allow us to output to stdout any logging produced by SimpleJwtAuth (in production that will be warns and errors).

Wrap the configuration block in a Rails `to_prepare` so the logger is already initialised by the time we are assigning it.