Allow apply to call the return application endpoint
In locking down the different endpoints, I forgot Apply make use of the return endpoint in the developer tools (non-prod) as a quick way to mark an application as returned without having to run Review locally.

This is of great help during development of the IoJ/Means passporting and edge cases as allow for a very quick feedback loop (everything can be done from Apply without Review involvement).
Also helpful for quick tests on staging.

For the time being until we don't need this anymore, enable this endpoint again to be called by Apply.
Before release, we can turn this off again, or allow it only on non-prod envs (localhost/staging).