CRIMAP-372 Add metabase manifests
As Metabase, at least initially, is going to consume the Datastore database (RDS), it makes sense to have the deployment manifests together in this repo.

Metabase manifests will not be applied automatically as part of any datastore deploy. Again, this is just an interim place to keep them versioned. The apply for now can be done manually if something needs to change.

I considered creating a separate repo just for this, but I don't think we have yet a clear vision of how to organise this.
Potentially, if we end up having an intermediate "analytics proxy service", this might sit on top of that, and we could move the manifests there.