Add a subdomain for metabase (#98)
Same as we have with `staging` subdomains. This ensures a more predictable and easy to remember URL.

For now it is a subdomain on apply, as datastore will not have soon public ingress, and review seemed like a less obvious choice.

If in the future we have a separate service for analytics, we can change it easily.