CRIMAP-383 Expose `application_type`
The attribute `application_type` was previously not propagated from Apply not exposed in the JSON. It is now exposed for normal applications and also for MAAT applications.

It uses the latest schemas version that incorporates this attribute.

As part of this work, I've tried to simplify a bit the application entities (ours, and MAAT one), so they use an inheritance approach as most of the code is identical. This should be flexible enough to cope with further changes in the future.