CRIMAP-385 Remove dependency on unneeded fixtures
This is part of an overall work to rationalise and remove any unneeded, unused, or superfluous fixtures, as these grow a bit out of hand for a while for no good reason.

On Datastore I've detected a couple superfluous fixtures: application_completed.json and application_invalid.json

These have been replaced by the overarching `application.json`, merging or tweaking the JSON where necessary for the specific test scenario.