CRIMAP-387 Return list of pruned applications
This uses the latest version of the schemas gem (v3.0.0) that supports a new type of schema/struct for a pruned application, reducing drastically the number of attributes returned in the `GET /applications` endpoint.

A new entity has been created to represent this pruned version of the applications, so it is independent of the "main" entity used in other endpoints.

Also, this endpoint is only used on Apply currently, to build the dashboard and paginate through applications, so there should be no impact on any other consumer (Review or MAAT).

A PR on Apply is also require along with this work, to be deployed together.