CRIMAP-402 Initial work for redacting PII details
NOTE: this is not fully functional but to the best of my efforts it will not break anything existing on the datastore. Current applications will stay unredacted, and only new submissions will be redacted.

However the `first_name` and `last_name` redacting is temporarily disabled because the `searchable_text` depends on these. The idea is to move this attribute and also `applicant_first_name` and `applicant_last_name`) to the new table `personally_identifiable_details` so they keep working.

Also, some work is needed to avoid (if at all possible, due to certain limitations) N+1 queries. This will be something for a follow-up PR but first we need to solve the `searchable_text` issue.