CRIMAP-402 Redacting PII details v2
Please for full context refer to existing open PR #106 as this is an iteration over that one.

This approach is quite similar (in fact I've been able to reuse almost all code with minor to no changes) with the difference the redacted payload (in its entirety, not just the redacted attributes) is stored in a separate table named `redacted_crime_applications`.

The original, unredacted payload, thus reside untouched in the original table `crime_applications`.

The advantage is, as we have it already unredacted, there is no join or any operation on reading back these applications (thus we eliminate N+1 issues).
The redacted table can then be used purely for analytics like metabase, by granting access to this table (instead of the "unredacted" one).

Also, no need to make changes to the `searchable_text` and other virtual attributes as these remain in the main, unredacted table like before.

The only disadvantage might be slightly more work to do for "right to forget", and possibly Metabase might lose access to some top level `crime_applications` table attributes (`returned_at`, `reviewed_at`, `offence_class`...)