Add `office_code` virtual attribute
On Apply all the listings in the dashboard are scoped by `status`, but also by `office_code`.

This should make these queries much more performant once we have hundreds or thousand of applications.

As the only place where these scopes are needed right now is in the `ListApplications`, I've made it explicit. If in the future we find ourselves using these scopes in other places, we can always add them back to the model.