Store return reason in the redacted metadata
The only bit of information that we need to perform analytics is the return reason. We don't need the free text `details`, and also this may contain PII.

As this is a separate table (`return_details`) we gave access to metabase, but we can remove access with this change.

Another alternative solution would be to refactor the `return_details` table so their attributes are part of the main `crime_application` table (maybe a jsonb or just first level attrs) and remove the table as no longer needed. It holds very little information as it is now.
But this involves more work and thinking as it is used in the entities, etc.