Relax some prometheus alerts (#115)
Similar changes we did to Apply. Refer to PR
https://github.com/ministryofjustice/laa-apply-for-criminal-legal-aid/pull/390

Note: on Apply error alerts is set to more than 5 errors in 5 mins. Here, given the datastore is a key, central component, better to err on the cautious side, so reducing to 3 errors in 5 mins. We can keep tweaking it depending how it goes.