Set metabase modsec allowed verbs (#117)
For some reason modsec was blocking PUT requests. I've also added DELETE as it is probably needed.
Basically we want to allow all verb methods as all are required.

Interestingly on Apply/Review this is not an issue (so far) but again we have to keep an eye on modsec logs.