CRIMAP-373 Implement prometheus exporter
This piggy backs on all work/discovery done firstly within the Apply service. It is identical, I've just not added any custom metrics collector as we don't know yet what we could measure (if anything).

One note tho, as this is a Grape application, and there are no classic controllers and actions like in a Rails app, sadly this gem does not seem to be prepared to cope with it, so metrics about http requests will all have the same tags `action="other", controller="other"`

It could be improved in the future.