Better grafana reports for Grape API endpoints (#128)
Currently all API endpoints have controller "other" and action "other", and most of the grafana dashboards work with these controller/action.

To make it more meaningful, the same we had custom labels, we now have default labels for grape API endpoints, where the controller is `(api)` and the action is whatever the endpoint was. This allows for better metrics when using unified Grafana dashboards like golden signals.