CRIMAP-356 Deauthorise `return` endpoint to apply consumer
Apply uses this endpoint to trigger application returns from the developer tools footer menu. It is a quick way to test the return journey without needing Review up and running.

However this is not needed in production, and thus this is now env-dependant and the endpoint will not be available to `crime-apply` consumer in production namespace, reducing risk.

This is the only endpoint that requires this differentiation.