CRIMAP-436 Expose `ioj_bypass` in MAAT json
Necessary changes to go along with the new schemas gem version.

As discussed, the `interests_of_justice` attribute is no longer returned in the MAAT json, and is replaced with a new `ioj_bypass` boolean, whether we captured interests of justice reasons from the provider, or not.

It simplifies business logic for MAAT injection.

Apply / Review services are not affected by this changes and there is no need to bump the schemas gem just for this.