CRIMAP-464 Unredact legal rep name attributes
This is part of a wider investigation Chris Purvis is performing.

So far these 2 need to be unredacted as will be used for reporting purposes. This is information that is required by contract managers.

Clarification on DoB is required, but the person in charge of that is on holidays until Monday.