CRIMAP-525 Implement first few document endpoints
This is a first part of a series of several PRs, not only on datastore but also counterpart API gem, to operate with documents in an S3 bucket.

As part of this first PR a little bit of ground preparation has been done and 2 endpoints implemented, to obtain presigned URLs for uploads and downloads.

I will raise a similar PR on the datastore API gem to be able to call these endpoints.