CRIMAP-530 S3 local development setup
In order to ease development and testing of anything related to evidence upload (S3 buckets), and given very soon we will not be able to use long-lived credentials from staging, this is an initial basic local setup of an S3-like localstack container that can be used to mimic the S3 API.

The setup should be minimal, just run the container and ensure you remove or comment out any previous S3 env variables in your `.env.development.local` if you added any.

Note for now there is no persistence so once you kill this container any uploaded files are gone. This will be improved if possible later on (thus the `shutdown.sh` hook).

Unfortunately persistence is a paid feature needing a Pro localstack API key, but it seems it may be possible to do some work around.
If not, then we could make a case about getting a pro account, if we feel this may be valuable for us.

In a separate PR will try to also take a similar approach to replace elasticmq and SNS faker with localstack if possible.