Use LocalStack also for SNS messaging
Cleaned up old elasticmq references as it is no longer needed if we can do the same and better with LocalStack.

Better and more consistent handling/naming of ENV variables.

Updated README.

NOTE: changes are needed on Review in order for the SNS notifications endpoint to work and to confirm the subscription, etc. This will be discussed separately, if required.