Use LocalStack also for SNS messaging (#144)
* Use LocalStack also for SNS messaging

Cleaned up old elasticmq references as it is no longer needed if we can do the same and better with LocalStack.

Better and more consistent handling/naming of ENV variables.

Updated README.

NOTE: changes are needed on Review in order for the SNS notifications endpoint to work and to confirm the subscription, etc. This will be discussed separately, if required.

* Simplifying some env variables

And making sure Prometheus exporter, if enabled, can be configured in a different port, otherwise all containers will use same port and will be a clash.