Allow MAAT adapter dev and uat to use staging
New API secrets for MAAT `dev` and `uat` have been injected in their corresponding namespaces and these changes are needed to pick them from the ENV and use them with the JWT authentication.

Their previous solution was to manually hardcode the staging secret so this is an improvement until they move to use a mock sometime in the future.