Attempt to fix workflow ECR_URL
The previously used `secrets.ECR_URL` is now removed as part of ECR migration to short-term creds so is not available anymore to use in the `./.github/actions/deploy` action.

However we can use reference job outputs in other job steps.

This is a first attempt I'm not 100% if this will work but worth a try.

NOTE: in order to reduce feedback I'm temporarily enabling the build/deploy in one go without merging to main.