CRIMAP-600 Expose first court hearing attributes
Needed to render completed application and for rehydration purposes.

Unexposing these details from MAAT json on purpose for the time being.

Please do not merge until Apply/Review validate the changes and make sure it all works for multiple combinations: attributes are not present in the JSON, attributes are present but nil, etc.